Utility/Blob, Utility/MemSet moved to Utility/Memory folder
