+++ src/Engine/Graphics/Direct3D/Render.cpp
@@ -41,7 +41,7 @@
 
 #include "Io/Mouse.h"
 
-#include "Utility/Memory.h"
+#include "Utility/Memory/MemSet.h"
 #include "Utility/Math/TrigLut.h"
 #include "Utility/Random/Random.h"
