+++ src/Engine/Tables/PlayerFrameTable.h
@@ -2,7 +2,7 @@
 
 #include "Engine/Objects/Player.h"
 
-#include "Utility/Blob.h"
+#include "Utility/Memory/Blob.h"
 
 /*   46 */
 #pragma pack(push, 1)
