+++ src/Engine/Tables/TileFrameTable.h
@@ -4,7 +4,7 @@
 
 #include "Engine/AssetsManager.h"
 
-#include "Utility/Blob.h"
+#include "Utility/Memory/Blob.h"
 
 /*  327 */
 enum TILE_DESC_FLAGS {
