+++ src/GUI/UI/Books/JournalBook.cpp
@@ -19,7 +19,7 @@
 
 #include "Media/Audio/AudioPlayer.h"
 
-#include "Utility/Memory.h"
+#include "Utility/Memory/MemSet.h"
 
 Image *ui_book_journal_background = nullptr;
