+++ src/Utility/Memory/Blob.cpp
